[ADD] Adding of first file system
This if first file system management, it manage files with depth, it can only create files for now.

General (./):
	- ./Makefile: Adding of files for compilation
	- ./kernel(.c/.h): add of initialization for file system

Daniel-tty (./src/GUI/tools/tty/):
	- ./buffer.c: discard warning
	- ./commands.c: discard warning

Drivers (./src/drivers/mouse/):
	- ./mouse.c: discard_warning

Systems (./src/systems/):
	- ./files.h: creating of file system api
	- ./files/create(.c/.h): initialization of files and of firsts function of file creating
	- ./files/init(.c/.h): initialization of file system
	- ./files/intern.h: initialization of intern variable

Utils (./src/utils/):
	- ./string.h: changing of strlen return value from int to size_t
	- ./string/split.h: warning discard
	- ./string/strlen(.c/.h): changing return value to size_t

+++ src/GUI/tools/tty/buffer.c
@@ -37,7 +37,7 @@ static int want_delete_prompt()
     int have_prompt = 0;
 
     for (size_t i = 0; this.prompt_index[i] != -1; i++) {
-        if (this.prompt_index[i] == this.cy) {
+        if ((size_t)this.prompt_index[i] == this.cy) {
             have_prompt = 1;
             break;
         }
