+++ src/utils/string/strlen.c
@@ -8,9 +8,9 @@
  * @param s : the string to be measured
  * @return int : the length of the string
  */
-int strlen(const char *s)
+size_t strlen(const char *s)
 {
-    int len = 0;
+    size_t len = 0;
 
     for (; s[len]; len++);
     return (len);
